Allow to use this TextMate bundle with any 2.x.y version of Ruby
If the `ruby20` is specified to run the commands in this bundle and
the machine running TextMate has Ruby 2.x.y installed as the system
Ruby, the system Ruby will be used. This basically means that any Ruby
2.x.y version is allowed. Gems are installed in a directory with the
same name as the Ruby ABI version, since this can change for minor
version. Because of that, we need to use a custom location for the
gems that works regardless of the Ruby ABI version. This should work
fine since no C extensions are used by any of the gems used by this
bundle.